v0.4.3-beta release
This release brins MANY new stuff.

First, MDroid now parses index-v1.jar/index-v1.json to get localized data (usefull for later once the app supports i18n).

Then, apps now can show... SCREENSHOTS !

And maybe the most important of all, removing that shit XML gives a big boost to the app's performances. (in reallity it's still used for backward compatibility with repos that didn't updated to fdroidserver v1)

And maybe one the of the most important change is the removal of `react-native-cached-image` in favor of `react-native-fast-image`. A Glid/SDImage wrapper for react-native. So the image caching is done the native side, and that allows the main UI thread to not be busy anymore on startup !

ENJOOOOOOOOOOOOOY !!!!