BIG CODE REFACTORING GUYS!
Closes #14 ! Hell yeah.

This commit is an huge one. Yep it is. God it IS! I refactored almost all the app's files, lol.

This should gives the app a big performances boost. Thanks to the react-native-debugger Profiler for the help!

It was a nightmare, and I still have to refactor that shit AppDetails screen, but releasing this essentially for testing it in production mode and see if anything bad happens. :yum:

Anyway, enjoy!